Factor out common layer
